split direct build graph validation fixtures
